+++ ibllib/io/extractors/camera.py
@@ -7,9 +7,9 @@ from functools import partial
 import cv2
 import numpy as np
 import matplotlib.pyplot as plt
+from iblutil.util import range_str
 
 import ibllib.dsp.utils as dsp
-from iblutil.util import range_str
 from ibllib.plots import squares, vertical_lines
 from ibllib.io.video import assert_valid_label, VideoStreamer
 from iblutil.numerical import within_ranges