+++ ibllib/pipes/ephys_preprocessing.py
@@ -10,6 +10,7 @@ import numpy as np
 import pandas as pd
 import one.alf.io as alfio
 
+from ibllib.misc import check_nvidia_driver
 from ibllib.ephys import ephysqc, spikes, sync_probes
 from ibllib.io import ffmpeg, spikeglx
 from ibllib.io.video import label_from_path
@@ -19,7 +20,7 @@ from ibllib.pipes.training_preprocessing import TrainingRegisterRaw as EphysRegi
 from ibllib.qc.task_extractors import TaskQCExtractor
 from ibllib.qc.task_metrics import TaskQC
 from ibllib.qc.camera import run_all_qc as run_camera_qc
-from ibllib.dsp import rms, voltage
+from ibllib.dsp import rms
 
 _logger = logging.getLogger("ibllib")
 
@@ -80,14 +81,19 @@ class EphysAudio(tasks.Task):
         return [output_file]
 
 
-class SpikeSorting_KS2_Matlab(tasks.Task):
+class SpikeSorting(tasks.Task):
     """
-    Computes raw electrophysiology QC
+    Pykilosort 2.5 pipeline
     """
     gpu = 1
     io_charge = 70  # this jobs reads raw ap files
     priority = 60
     level = 1  # this job doesn't depend on anything
+    SHELL_SCRIPT = Path.home().joinpath(
+        "Documents/PYTHON/iblscripts/deploy/serverpc/kilosort2/run_pykilosort.sh"
+    )
+    SPIKE_SORTER_NAME = 'pykilosort'
+    PYKILOSORT_REPO = Path.home().joinpath('Documents/PYTHON/SPIKE_SORTING/pykilosort')
 
     @staticmethod
     def _sample2v(ap_file):
@@ -96,40 +102,55 @@ class SpikeSorting_KS2_Matlab(tasks.Task):
         return s2v["ap"][0]
 
     @staticmethod
-    def _fetch_ks2_commit_hash():
-        command2run = "git --git-dir ~/Documents/MATLAB/Kilosort2/.git rev-parse --verify HEAD"
+    def _fetch_pykilosort_version(repo_path):
+        init_file = Path(repo_path).joinpath('pykilosort', '__init__.py')
+        version = SpikeSorting._fetch_ks2_commit_hash(repo_path)  # default
+        try:
+            with open(init_file) as fid:
+                lines = fid.readlines()
+                for line in lines:
+                    if line.startswith("__version__ = "):
+                        version = line.split('=')[-1].strip().replace('"', '').replace("'", '')
+        except Exception:
+            pass
+        return f"pykilosort_{version}"
+
+    @staticmethod
+    def _fetch_ks2_commit_hash(repo_path):
+        command2run = f"git --git-dir {repo_path}/.git rev-parse --verify HEAD"
         process = subprocess.Popen(
             command2run, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
         )
         info, error = process.communicate()
         if process.returncode != 0:
             _logger.error(
-                f"Can't fetch matlab ks2 commit hash, will still attempt to run \n"
+                f"Can't fetch pykilsort commit hash, will still attempt to run \n"
                 f"Error: {error.decode('utf-8')}"
             )
             return ""
         return info.decode("utf-8").strip()
 
-    def _run_ks2(self, ap_file):
-        """
+    def _run_pykilosort(self, ap_file):
+        f"""
         Runs the ks2 matlab spike sorting for one probe dataset
-        the spike sorting output can either be with the probe (<1.5.5) or in the
-        session_path/spike_sorters/ks2_matlab/probeXX folder
+        the raw spike sorting output can either be with the probe (<1.5.5) or in the
+        session_path/spike_sorters/{self.SPIKE_SORTER_NAME}/probeXX folder
         :return: path of the folder containing ks2 spike sorting output
         """
-        label = ap_file.parts[-2]
-        if ap_file.parent.joinpath("spike_sorting_ks2.log").exists():
-            _logger.info(f"Already ran: spike_sorting_ks2.log found for {ap_file}, skipping.")
+        self.version = self._fetch_pykilosort_version(self.PYKILOSORT_REPO)
+        label = ap_file.parts[-2]  # this is usually the probe name
+        if ap_file.parent.joinpath(f"spike_sorting_{self.SPIKE_SORTER_NAME}.log").exists():
+            _logger.info(f"Already ran: spike_sorting_{self.SPIKE_SORTER_NAME}.log"
+                         f" found for {ap_file}, skipping.")
             return ap_file.parent
-        ks2_dir = self.session_path.joinpath("spike_sorters", "ks2_matlab", label)
-        if ks2_dir.joinpath("spike_sorting_ks2.log").exists():
-            _logger.info(f"Already ran: spike_sorting_ks2.log found in {ks2_dir}, skipping.")
-            return ks2_dir
+        sorter_dir = self.session_path.joinpath("spike_sorters", self.SPIKE_SORTER_NAME, label)
+        print(sorter_dir.joinpath(f"spike_sorting_{self.SPIKE_SORTER_NAME}.log"))
+        if sorter_dir.joinpath(f"spike_sorting_{self.SPIKE_SORTER_NAME}.log").exists():
+            _logger.info(f"Already ran: spike_sorting_{self.SPIKE_SORTER_NAME}.log"
+                         f" found in {sorter_dir}, skipping.")
+            return sorter_dir
         # get the scratch drive from the shell script
-        SHELL_SCRIPT = Path.home().joinpath(
-            "Documents/PYTHON/iblscripts/deploy/serverpc/kilosort2/task_ks2_matlab.sh"
-        )
-        with open(SHELL_SCRIPT) as fid:
+        with open(self.SHELL_SCRIPT) as fid:
             lines = fid.readlines()
         line = [line for line in lines if line.startswith("SCRATCH_DRIVE=")][0]
         m = re.search(r"\=(.*?)(\#|\n)", line)[0]
@@ -137,33 +158,19 @@ class SpikeSorting_KS2_Matlab(tasks.Task):
         assert scratch_drive.exists()
 
         # clean up and create directory, this also checks write permissions
-        # scratch dir has the following shape: ks2m/ZM_3003_2020-07-29_001_probe00
+        # temp_dir has the following shape: pykilosort/ZM_3003_2020-07-29_001_probe00
         # first makes sure the tmp dir is clean
-        shutil.rmtree(scratch_drive.joinpath("ks2m"), ignore_errors=True)
-        scratch_dir = scratch_drive.joinpath(
-            "ks2m", "_".join(list(self.session_path.parts[-3:]) + [label])
+        shutil.rmtree(scratch_drive.joinpath(self.SPIKE_SORTER_NAME), ignore_errors=True)
+        temp_dir = scratch_drive.joinpath(
+            self.SPIKE_SORTER_NAME, "_".join(list(self.session_path.parts[-3:]) + [label])
         )
-        if scratch_dir.exists():
-            shutil.rmtree(scratch_dir, ignore_errors=True)
-        scratch_dir.mkdir(parents=True, exist_ok=True)
+        if temp_dir.exists():  # hmmm this has to be decided, we may want to restart ?
+            # But failed sessions may then clog the scratch dir and have users run out of space
+            shutil.rmtree(temp_dir, ignore_errors=True)
+        temp_dir.mkdir(parents=True, exist_ok=True)
 
-        # check the nvidia status before decompressing so as to fail quicker
-        process = subprocess.Popen('nvidia-smi', shell=True, stdout=subprocess.PIPE,
-                                   stderr=subprocess.PIPE, executable="/bin/bash")
-        info, error = process.communicate()
-        if process.returncode != 0:
-            raise RuntimeError(f"Nvida drivers not ready. \n {error.decode('utf-8')}")
-        _logger.info("nvidia-smi command successful")
-
-        # decompresses using mtscomp
-        tmp_ap_file = scratch_dir.joinpath(ap_file.name).with_suffix(".bin")
-
-        voltage.decompress_destripe_cbin(sr=ap_file, output_file=tmp_ap_file)
-        # mtscomp.decompress(cdata=ap_file, out=tmp_ap_file)
-
-        # run matlab spike sorting: with R2019a, it would be much easier to run with
-        # -batch option as matlab errors are redirected to stderr automatically
-        command2run = f"{SHELL_SCRIPT} {scratch_dir}"
+        check_nvidia_driver()
+        command2run = f"{self.SHELL_SCRIPT} {ap_file} {temp_dir}"
         _logger.info(command2run)
         process = subprocess.Popen(
             command2run,
@@ -174,19 +181,20 @@ class SpikeSorting_KS2_Matlab(tasks.Task):
         )
         info, error = process.communicate()
         info_str = info.decode("utf-8").strip()
+        _logger.info(info_str)
         if process.returncode != 0:
-            raise RuntimeError(error.decode("utf-8"))
-        elif "run_ks2_ibl.m failed" in info_str:
-            _logger.info(info_str)
-            raise RuntimeError("Matlab error ks2 log below:")
+            error_str = error.decode("utf-8").strip()
+            # try and get the kilosort log if any
+            for log_file in temp_dir.rglob('*_kilosort.log'):
+                with open(log_file) as fid:
+                    log = fid.read()
+                    _logger.error(log)
+                break
+            raise RuntimeError(f"{self.SPIKE_SORTER_NAME} {info_str}, {error_str}")
 
-        # clean up and copy: output to session/spike_sorters/ks2_matlab/probeXX (ks2_dir)
-        tmp_ap_file.unlink()  # remove the uncompressed temp binary file
-        scratch_dir.joinpath("temp_wh.dat").unlink()  # remove the memmapped pre-processed file
-        shutil.move(scratch_dir, ks2_dir)
-
-        self.version = self._fetch_ks2_commit_hash()
-        return ks2_dir
+        shutil.copytree(temp_dir.joinpath('output'), sorter_dir, dirs_exist_ok=True)
+        shutil.rmtree(temp_dir, ignore_errors=True)
+        return sorter_dir
 
     def _run(self, overwrite=False):
         """
@@ -202,8 +210,9 @@ class SpikeSorting_KS2_Matlab(tasks.Task):
         out_files = []
         for ap_file, label in ap_files:
             try:
-                ks2_dir = self._run_ks2(ap_file)  # runs ks2, skips if it already ran
-                probe_out_path = self.session_path.joinpath("alf", label)
+                ks2_dir = self._run_pykilosort(ap_file)  # runs ks2, skips if it already ran
+                probe_out_path = self.session_path.joinpath("alf", label, self.SPIKE_SORTER_NAME)
+                shutil.rmtree(probe_out_path, ignore_errors=True)
                 probe_out_path.mkdir(parents=True, exist_ok=True)
                 spikes.ks2_to_alf(
                     ks2_dir,
@@ -212,12 +221,17 @@ class SpikeSorting_KS2_Matlab(tasks.Task):
                     bin_file=ap_file,
                     ampfactor=self._sample2v(ap_file),
                 )
+                logfile = ks2_dir.joinpath(f"spike_sorting_{self.SPIKE_SORTER_NAME}.log")
+                if logfile.exists():
+                    shutil.copyfile(logfile, probe_out_path.joinpath(
+                        f"_ibl_log.info_{self.SPIKE_SORTER_NAME}.log"))
                 out, _ = spikes.sync_spike_sorting(ap_file=ap_file, out_path=probe_out_path)
                 out_files.extend(out)
                 # convert ks2_output into tar file and also register
                 # Make this in case spike sorting is in old raw_ephys_data folders, for new
                 # sessions it should already exist
-                tar_dir = self.session_path.joinpath('spike_sorters', 'ks2_matlab', label)
+                tar_dir = self.session_path.joinpath(
+                    'spike_sorters', self.SPIKE_SORTER_NAME, label)
                 tar_dir.mkdir(parents=True, exist_ok=True)
                 out = spikes.ks2_to_tar(ks2_dir, tar_dir)
                 out_files.extend(out)
@@ -448,7 +462,7 @@ class EphysExtractionPipeline(tasks.Pipeline):
         tasks["EphysAudio"] = EphysAudio(self.session_path)
         tasks["EphysMtscomp"] = EphysMtscomp(self.session_path)
         # level 1
-        tasks["SpikeSorting"] = SpikeSorting_KS2_Matlab(
+        tasks["SpikeSorting"] = SpikeSorting(
             self.session_path, parents=[tasks["EphysMtscomp"], tasks["EphysPulses"]])
         tasks["EphysVideoCompress"] = EphysVideoCompress(
             self.session_path, parents=[tasks["EphysPulses"]])