+++ oneibl/__init__.py
@@ -1,1 +1,0 @@
-"""IBL pipeline-specific ONE functions"""