+++ release_notes.md
@@ -1,3 +1,19 @@
+## Release Notes 2.0.0
+### Release Notes 2.0.1 2021-08-07
+- pykilosort error handling
+### Release Notes 2.0.2 2021-08-31
+- passive extraction robust to frame2ttl flickers
+### Release Notes 2.0.3 2021-09-03
+- pykilosort bugfix after low yield results
+### Release Notes 2.0.4 2021-09-10
+- ephys trials extraction when audio FPGA starts on up state
+### Release Notes 2.0.5 2021-03-13
+- pykilosort pipeline: output correct version number / fix log file name
+
+### Release Notes 2.0.0 2021-08-04
+- ONE2 released on the master branch
+- Pykilosort is the new default spike sorter in the pipeline
+
 ## Release Notes 1.12 (not released yet)
 ### Release Notes 1.12.0
 - oneibl and alf now deprecated; moved to separate repository